Unify initialization tools for CPU and openCL
