+++ initializer/Cargo.toml
@@ -9,4 +9,7 @@ edition = "2021"
 base64 = "0.21.0"
 clap = { version = "4.2.4", features = ["derive"] }
 post-rs = { path = "../" }
+scrypt-ocl = { path = "../scrypt-ocl" }
 rayon = "1.7.0"
+eyre = "0.6.8"
+rand = "0.8.5"
